+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Arista EOS Role for Ansible
 
 - adds additional parameters to eos_config
 - fixes issue with honoring enablepwd if specified
+- adds vlan argument to eos_vxlan_vtep
 
 ## 1.0.0, 5/4/2015